Written a lot more eval functions. Not tested.
def, set, & var are very similar things. They are also very simple, it would be interesting to eval the chain of `def` when it is created not when evald.
Calling a list or dict acts like it take one argument 'x' it then looks up that value in the data.