Created boolean type. Added more built-in functions.
There is now a global environment that currently has just two elements, true and false. We have the basic maths operations in a new file for built in functions.

It wont take long to add the rest of the basic functions here. The most important ones are probably `equal` and `for`.

I basically need to write some code in this language now and see what works and what doesn't. I could either do this through testing or by making a program, I think writing a lexer for it would be fun.